+++ app/src/main/java/com/android/quizcafe/core/data/model/quizbook/response/CategoryResponseDto.kt
@@ -1,6 +1,6 @@
 package com.android.quizcafe.core.data.model.quizbook.response
 
-import com.android.quizcafe.core.domain.model.quizbook.reponse.Category
+import com.android.quizcafe.core.domain.model.quizbook.response.Category
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
