+++ app/src/main/java/com/android/quizcafe/core/domain/model/quizbook/response/QuizBookDetail.kt
@@ -1,4 +1,4 @@
-package com.android.quizcafe.core.domain.model.quizbook.reponse
+package com.android.quizcafe.core.domain.model.quizbook.response
 
 // TODO: API Response에 맞게 수정
 data class QuizBookDetail(
