+++ app/src/main/java/com/android/quizcafe/core/domain/repository/QuizBookRepository.kt
@@ -1,9 +1,9 @@
 package com.android.quizcafe.core.domain.repository
 
 import com.android.quizcafe.core.domain.model.Resource
-import com.android.quizcafe.core.domain.model.quizbook.reponse.Category
-import com.android.quizcafe.core.domain.model.quizbook.reponse.QuizBook
-import com.android.quizcafe.core.domain.model.quizbook.reponse.QuizBookDetail
+import com.android.quizcafe.core.domain.model.quizbook.response.Category
+import com.android.quizcafe.core.domain.model.quizbook.response.QuizBook
+import com.android.quizcafe.core.domain.model.quizbook.response.QuizBookDetail
 import com.android.quizcafe.core.domain.model.quizbook.request.CategoryRequest
 import kotlinx.coroutines.flow.Flow
