+++ app/src/main/java/com/android/quizcafe/core/domain/usecase/quizbook/GetCategoryListUseCase.kt
@@ -1,7 +1,7 @@
 package com.android.quizcafe.core.domain.usecase.quizbook
 
 import com.android.quizcafe.core.domain.model.Resource
-import com.android.quizcafe.core.domain.model.quizbook.reponse.Category
+import com.android.quizcafe.core.domain.model.quizbook.response.Category
 import com.android.quizcafe.core.domain.model.quizbook.request.CategoryRequest
 import com.android.quizcafe.core.domain.repository.QuizBookRepository
 import kotlinx.coroutines.flow.Flow
