+++ app/src/main/java/com/android/quizcafe/core/domain/usecase/quizbook/GetQuizBookDetailUseCase.kt
@@ -1,7 +1,7 @@
 package com.android.quizcafe.core.domain.usecase.quizbook
 
 import com.android.quizcafe.core.domain.model.Resource
-import com.android.quizcafe.core.domain.model.quizbook.reponse.QuizBookDetail
+import com.android.quizcafe.core.domain.model.quizbook.response.QuizBookDetail
 import com.android.quizcafe.core.domain.repository.QuizBookRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
