+++ app/src/main/java/com/android/quizcafe/core/domain/usecase/quizbook/GetQuizBookListUseCase.kt
@@ -1,7 +1,7 @@
 package com.android.quizcafe.core.domain.usecase.quizbook
 
 import com.android.quizcafe.core.domain.model.Resource
-import com.android.quizcafe.core.domain.model.quizbook.reponse.QuizBook
+import com.android.quizcafe.core.domain.model.quizbook.response.QuizBook
 import com.android.quizcafe.core.domain.repository.QuizBookRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
